added loading results function for qutip and fixed bugs in dynamiqs
