+++ tests/instruments/emulator/test_gpu.py
@@ -1,9 +1,6 @@
 """Testing emulator GPU configuration."""
 
 import importlib.util
-import os
-import subprocess
-import sys
 
 import numpy as np
 import pytest
@@ -29,40 +26,13 @@ def test_qutip_gpu_unknown_dtype():
         QutipEngine(device="gpu", gpu_dtype="unknown").destroy(2)
 
 
-@pytest.mark.parametrize(
-    "gpu_dtype,plugin", [("jax", "qutip_jax"), ("cupyd", "qutip_cupy")]
-)
+@pytest.mark.parametrize("gpu_dtype,plugin", [("jax", "qutip_jax")])
 def test_qutip_gpu_requires_plugin(gpu_dtype, plugin):
     if importlib.util.find_spec(plugin) is not None:
         pytest.skip(f"{plugin} is installed in this environment.")
 
     with pytest.raises(ImportError, match=plugin.replace("_", "-")):
         QutipEngine(device="gpu", gpu_dtype=gpu_dtype).destroy(2)
-
-
-def test_qutip_gpu_jax_requires_jax_gpu_backend():
-    if importlib.util.find_spec("qutip_jax") is None:
-        pytest.skip("qutip_jax is not installed in this environment.")
-
-    env = os.environ.copy()
-    env["JAX_PLATFORMS"] = "cpu"
-    result = subprocess.run(
-        [
-            sys.executable,
-            "-c",
-            (
-                "from qibolab._core.instruments.emulator.engine import QutipEngine; "
-                "QutipEngine(device='gpu', gpu_dtype='jax').destroy(2)"
-            ),
-        ],
-        env=env,
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-
-    assert result.returncode != 0
-    assert "CUDA-enabled JAX backend" in result.stderr
 
 
 @pytest.mark.parametrize("gpu_dtype", ["jax", "cupyd"])
